Changed to EntityManager.

+++ src/main/java/com/biit/persistence/dao/IGenericDao.java
@@ -32,7 +32,7 @@ public interface IGenericDao<T> {
 	 * 
 	 * @return
 	 */
-	int getRowCount();
+	Long getRowCount();
 
 	/**
 	 * Gets one element by id.
